Change trajectories to use ints instead of shorts

+++ src/kbmod/search/common.h
@@ -42,10 +42,10 @@ struct Trajectory {
     // Est. Flux
     float flux = 0.0;
     // Origin
-    short x = 0;
-    short y = 0;
+    int x = 0;
+    int y = 0;
     // Number of images summed
-    short obs_count;
+    int obs_count;
     // Whether the trajectory is valid. Used for on-GPU filtering.
     bool valid = true;
 
@@ -216,8 +216,8 @@ static void trajectory_bindings(py::module &m) {
                     [](py::tuple t) {  // __setstate__
                         if (t.size() != 8) throw std::runtime_error("Invalid state!");
                         tj trj = {t[0].cast<float>(), t[1].cast<float>(), t[2].cast<float>(),
-                                  t[3].cast<float>(), t[4].cast<short>(), t[5].cast<short>(),
-                                  t[6].cast<short>(), t[7].cast<bool>()};
+                                  t[3].cast<float>(), t[4].cast<int>(), t[5].cast<int>(),
+                                  t[6].cast<int>(), t[7].cast<bool>()};
                         return trj;
                     }));
 }
