+++ src/kbmod/search/stack_search.h
@@ -48,7 +48,7 @@ public:
 
     // The primary search functions
     void evaluate_single_trajectory(Trajectory& trj);
-    Trajectory search_linear_trajectory(short x, short y, float vx, float vy);
+    Trajectory search_linear_trajectory(int x, int y, float vx, float vy);
     void prepare_search(std::vector<Trajectory>& search_list, int min_observations);
     std::vector<Trajectory> search_single_batch();
     void search_batch();
